Remove changeState method
Removed change state to refactor several sections within in it inside the method moveToNextMachine. The Goal is to exttract several small methods out of this method to shorten it and to increase readability.
We also declared several variables assigning call chains to each of them to reduce the amount of long call chains in the method
Co-authored-by: Austin Robinson <robi1467@morris.umn.edu>